Updated git with Database_Store.php Began creating intermediate datalayer between frontend and backend 3/16/22 Updated test file to reflect the new Line Item factory 3/17/22 Updated test file and Line Item factory with initial validation logic 3/17/22 Tweaked the testing file formatting and minor changes to the  Line Item factory 3/17/22 After a few days of fighting with Apache & MYSQL had to rebuild Database_Store 3/24/22 Further adventures in SQL errors! 3/28/22 added ADT for strings and numbers added further data types cleaning up after the new types were introduced removing composer vendor folder from git updated Tax and Taxes classes More work on the Tax and Taxes classes. Ready to test and work the kinks out. Initial testing and refactoring complete. no immediate errors. moved Vendor folder to correct location updated main_test_2.php Got database connection working

+++ src/Utilities.php
@@ -21,11 +21,11 @@ class Utilities
        return $str;
     }
 
-    public static function random_price(): float{
-        //I want the format 0.00
-        $dollars = (float)mt_rand(0, 500);
-        $tens = (float)mt_rand(0, 9)/10;
-        $ones = (float)mt_rand(0, 9)/100;
+    public static function random_price(): int{
+        //I want the format in cents
+        $dollars = mt_rand(0, 500);
+        $tens = mt_rand(0, 9)/10;
+        $ones = mt_rand(0, 9)/100;
 
         return $dollars + $tens + $ones;
     }
